requirements.txt
cryptography
django
djangorestframework
django-cors-headers
drf-jwt
pyjwt
requests
python-dotenv